requirements.txt
streamlit
google-generativeai
GitPython
python-dotenv